backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-nmap==0.7.1
scapy==2.5.0
pydantic==2.6.0
python-multipart==0.0.7
jinja2==3.1.3
websockets==12.0
regex
